+++ src/main/java/net/sf/jabref/gui/preftabs/PreferencesDialog.java
@@ -15,14 +15,11 @@
  */
 package net.sf.jabref.gui.preftabs;
 
-import java.awt.BorderLayout;
-import java.awt.CardLayout;
-import java.awt.Component;
+import java.awt.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.io.File;
 import java.util.ArrayList;
-import java.util.Iterator;
 import java.util.List;
 
 import javax.swing.AbstractAction;
@@ -34,8 +31,6 @@ import javax.swing.JList;
 import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.ListSelectionModel;
-import javax.swing.event.ListSelectionEvent;
-import javax.swing.event.ListSelectionListener;
 
 import net.sf.jabref.*;
 import net.sf.jabref.exporter.ExportFormats;
@@ -70,33 +65,28 @@ public class PreferencesDialog extends JDialog {
 
     private final JabRefPreferences prefs;
 
-    private final JButton importPrefs = new JButton(Localization.lang("Import preferences"));
-    private final JButton exportPrefs = new JButton(Localization.lang("Export preferences"));
+    private final JButton importPreferences = new JButton(Localization.lang("Import preferences"));
+    private final JButton exportPreferences = new JButton(Localization.lang("Export preferences"));
+    private final JButton showPreferences = new JButton(Localization.lang("Show preferences"));
 
     private static final Log LOGGER = LogFactory.getLog(PreferencesDialog.class);
-
-
 
     public PreferencesDialog(JabRefFrame parent, JabRef jabRef) {
         super(parent, Localization.lang("JabRef preferences"), false);
         prefs = JabRefPreferences.getInstance();
         frame = parent;
 
-
         main = new JPanel();
-        JPanel upper = new JPanel();
+        JPanel mainPanel = new JPanel();
         JPanel lower = new JPanel();
 
         getContentPane().setLayout(new BorderLayout());
-        getContentPane().add(upper, BorderLayout.CENTER);
+        getContentPane().add(mainPanel, BorderLayout.CENTER);
         getContentPane().add(lower, BorderLayout.SOUTH);
 
         final CardLayout cardLayout = new CardLayout();
         main.setLayout(cardLayout);
 
-        // ----------------------------------------------------------------
-        // Add tabs to tabbed here. Remember, tabs must implement PrefsTab.
-        // ----------------------------------------------------------------
         List<PrefsTab> tabs = new ArrayList<>();
         tabs.add(new GeneralTab(prefs));
         tabs.add(new NetworkTab(prefs));
@@ -115,20 +105,13 @@ public class PreferencesDialog extends JDialog {
         tabs.add(new XmpPrefsTab());
         tabs.add(new AdvancedTab(prefs, jabRef));
 
-        Iterator<PrefsTab> prefTabs = tabs.iterator();
-        String[] names = new String[tabs.size()];
-        int index = 0;
+        // add all tabs
+        tabs.forEach(tab -> main.add((Component) tab, tab.getTabName()));
 
-        while (prefTabs.hasNext()) {
-            PrefsTab tab = prefTabs.next();
-            names[index] = tab.getTabName();
-            index++;
-            main.add((Component) tab, tab.getTabName());
-        }
+        mainPanel.setBorder(BorderFactory.createEtchedBorder());
 
-        upper.setBorder(BorderFactory.createEtchedBorder());
-
-        chooser = new JList<>(names);
+        String[] tabNames = tabs.stream().map(PrefsTab::getTabName).toArray(String[]::new);
+        chooser = new JList<>(tabNames);
         chooser.setBorder(BorderFactory.createEtchedBorder());
         // Set a prototype value to control the width of the list:
         chooser.setPrototypeCellValue("This should be wide enough");
@@ -137,29 +120,28 @@ public class PreferencesDialog extends JDialog {
 
         // Add the selection listener that will show the correct panel when
         // selection changes:
-        chooser.addListSelectionListener(new ListSelectionListener() {
-
-            @Override
-            public void valueChanged(ListSelectionEvent e) {
-                if (e.getValueIsAdjusting()) {
-                    return;
-                }
-                String o = chooser.getSelectedValue();
-                cardLayout.show(main, o);
+        chooser.addListSelectionListener(e -> {
+            if (e.getValueIsAdjusting()) {
+                return;
             }
+            String o = chooser.getSelectedValue();
+            cardLayout.show(main, o);
         });
 
-        JPanel one = new JPanel();
-        JPanel two = new JPanel();
-        one.setLayout(new BorderLayout());
-        two.setLayout(new BorderLayout());
-        one.add(chooser, BorderLayout.CENTER);
-        one.add(importPrefs, BorderLayout.SOUTH);
-        two.add(one, BorderLayout.CENTER);
-        two.add(exportPrefs, BorderLayout.SOUTH);
-        upper.setLayout(new BorderLayout());
-        upper.add(two, BorderLayout.WEST);
-        upper.add(main, BorderLayout.CENTER);
+
+        JPanel buttons = new JPanel();
+        buttons.setLayout(new GridLayout(3, 1));
+        buttons.add(importPreferences, 0);
+        buttons.add(exportPreferences, 1);
+        buttons.add(showPreferences, 2);
+
+        JPanel westPanel = new JPanel();
+        westPanel.setLayout(new BorderLayout());
+        westPanel.add(chooser, BorderLayout.CENTER);
+        westPanel.add(buttons, BorderLayout.SOUTH);
+        mainPanel.setLayout(new BorderLayout());
+        mainPanel.add(main, BorderLayout.CENTER);
+        mainPanel.add(westPanel, BorderLayout.WEST);
 
         JButton ok = new JButton(Localization.lang("OK"));
         JButton cancel = new JButton(Localization.lang("Cancel"));
@@ -177,9 +159,8 @@ public class PreferencesDialog extends JDialog {
         Util.bindCloseDialogKeyToCancelAction(this.getRootPane(), cancelAction);
 
         // Import and export actions:
-        exportPrefs.setToolTipText(Localization.lang("Export preferences to file"));
-        importPrefs.setToolTipText(Localization.lang("Import preferences from file"));
-        exportPrefs.addActionListener(new ActionListener() {
+        exportPreferences.setToolTipText(Localization.lang("Export preferences to file"));
+        exportPreferences.addActionListener(new ActionListener() {
 
             @Override
             public void actionPerformed(ActionEvent e) {
@@ -206,7 +187,8 @@ public class PreferencesDialog extends JDialog {
             }
         });
 
-        importPrefs.addActionListener(new ActionListener() {
+        importPreferences.setToolTipText(Localization.lang("Import preferences from file"));
+        importPreferences.addActionListener(new ActionListener() {
 
             @Override
             public void actionPerformed(ActionEvent e) {
@@ -230,6 +212,8 @@ public class PreferencesDialog extends JDialog {
             }
 
         });
+
+        showPreferences.addActionListener(e -> new JabRefPreferencesFilter.JabRefPreferencesDialog(new JabRefPreferencesFilter(Globals.prefs), frame).setVisible(true));
 
         setValues();